crypto-replicator/requirements.txt
# Crypto-Replicator Dependencies
# NIS2Shield Encrypted Twin Architecture

# Database
psycopg2-binary>=2.9.9

# Cryptography
cryptography>=41.0.0

# HTTP client
requests>=2.31.0

# Configuration
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
